add AddRectangle tool, refactor Pos to use Double, add rectOnDrag to EventHelper, add listener for adding children to editor pane, refactor select -listener, add preliminary delete on key pressed (wip)

+++ src/main/scala/ID/projects/Shape.scala
@@ -1,7 +1,7 @@
 package ID.projects
 import scalafx.scene.paint.Color
 
-trait Shape(private var _width: Double, private var _length: Double, private var _height: Double, private var _color: Color, private var _pos: Pos, val posOffset: (Int, Int, Int) = (0, 0, 0)):
+trait Shape(private var _width: Double, private var _length: Double, private var _height: Double, private var _color: Color, private var _pos: Pos, val posOffset: (Double, Double, Double) = (0, 0, 0)):
   def color = _color
   def width = _width
   def length = _length
@@ -27,11 +27,11 @@ trait Shape(private var _width: Double, private var _length: Double, private var
     scaleX(amount)
     scaleY(amount)
 
-class Rectangle(width: Int, length: Int, height: Int, color: Color, pos: Pos, posOffset: (Int, Int, Int)) extends Shape(length, width, height, color, pos, posOffset):
+class Rectangle(width: Int, length: Int, height: Int, color: Color, pos: Pos, posOffset: (Double, Double, Double)) extends Shape(length, width, height, color, pos, posOffset):
   def area = width * length
 
-class Oval(width: Int, length: Int, height: Int, color: Color, pos: Pos, posOffset: (Int, Int, Int)) extends Shape(length, width, height, color, pos, posOffset):
+class Oval(width: Int, length: Int, height: Int, color: Color, pos: Pos, posOffset: (Double, Double, Double)) extends Shape(length, width, height, color, pos, posOffset):
   def area = Math.PI * width * length
 
-class Triangle(width: Int, length: Int, height: Int, color: Color, pos: Pos, posOffset: (Int, Int, Int)) extends Shape(length, width, height, color, pos, posOffset):
+class Triangle(width: Int, length: Int, height: Int, color: Color, pos: Pos, posOffset: (Double, Double, Double)) extends Shape(length, width, height, color, pos, posOffset):
   def area = width * length / 2.0
